Rejig components.
We'll want each layer mapped from the toolkit and apidocs state with
differing child components, so rejig to make the layers easier to
create and the children state agnostic.
